Move collect_available_test_cases to check_test_cases.py
No behavior change.
